+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# version 0.2.0
+
+Added support for postgres types (enum and composite)
+
+# version 0.1.3
+
+Added 
+
+Changed  
+
+Fixed
+ - Omission of first function if function was schema's only item.
+
 # version 0.1.2
 
 Added 
@@ -13,7 +26,11 @@ Fixed
 # Unreleased
 
 [x] Make connection string configurable (user, host) [v0.1.2]
-[ ] Add export of types
+[/] Add export of types [v0.2.0, no range or box types]
+[ ] Add export of sequences
+[ ] Add export of table
+[ ] Add update/insert scripts 
+[ ] Add mode for querying through wait mode and an open file
 [ ] Add export of roles and privileges
 [ ] React to deletion of files (possibly also addition of files?)